chore(sched): Refactor `DurativeAction` and `Activity` to share most of their implementation.
